+++ apps/prairielearn/src/lib/res-locals.ts
@@ -13,19 +13,12 @@ import type { ResLocalsAssessment } from '../middlewares/selectAndAuthzAssessmen
 import type { ResLocalsAssessmentInstance } from '../middlewares/selectAndAuthzAssessmentInstance.js';
 import type { ResLocalsAssessmentQuestion } from '../middlewares/selectAndAuthzAssessmentQuestion.js';
 import type { ResLocalsInstanceQuestion } from '../middlewares/selectAndAuthzInstanceQuestion.js';
-import type {
-  ResLocalsInstructorQuestion,
-  ResLocalsInstructorQuestionSchema,
-} from '../middlewares/selectAndAuthzInstructorQuestion.js';
+import type { ResLocalsInstructorQuestion } from '../middlewares/selectAndAuthzInstructorQuestion.js';
 import type { ResLocalsCourseIssueCount } from '../middlewares/selectOpenIssueCount.js';
 
 import type { ResLocalsAuthnUser } from './authn.types.js';
 import type { ResLocalsConfig } from './config.js';
 import type { Course, CourseInstance } from './db-types.js';
-import type {
-  ResLocalsInstanceQuestionRender,
-  ResLocalsQuestionRender,
-} from './question-render.types.js';
 
 interface ResLocals extends ResLocalsAuthnUser, ResLocalsConfig, ResLocalsDate {
   __csrf_token: string;
@@ -43,34 +36,23 @@ interface ResLocalsForPageLookup {
   'instructor-instance-question': ResLocals &
     ResLocalsCourseIssueCount &
     ResLocalsCourseInstance &
-    ResLocalsInstructorQuestionSchema &
-    ResLocalsInstanceQuestion &
-    ResLocalsInstanceQuestionRender &
-    ResLocalsQuestionRender & {
-      questionRenderContext: 'manual_grading' | 'ai_grading';
+    ResLocalsInstructorQuestion &
+    ResLocalsInstanceQuestion & {
       navbarType: 'instructor';
     };
   'instructor-question': ResLocals &
     ResLocalsCourse &
     ResLocalsCourseIssueCount &
     Partial<ResLocalsCourseInstance> &
-    ResLocalsInstructorQuestion &
-    ResLocalsQuestionRender;
+    ResLocalsInstructorQuestion;
   'instructor-assessment-question': ResLocals &
     ResLocalsCourseIssueCount &
     ResLocalsCourseInstance &
     ResLocalsInstructorQuestion &
-    ResLocalsQuestionRender &
     ResLocalsAssessment &
     ResLocalsAssessmentQuestion;
-  'instance-question': ResLocals &
-    ResLocalsCourseInstance &
-    ResLocalsInstanceQuestion &
-    ResLocalsInstanceQuestionRender;
-  'assessment-question': ResLocals &
-    ResLocalsAssessment &
-    ResLocalsAssessmentQuestion &
-    ResLocalsInstanceQuestionRender;
+  'instance-question': ResLocals & ResLocalsCourseInstance & ResLocalsInstanceQuestion;
+  'assessment-question': ResLocals & ResLocalsAssessment & ResLocalsAssessmentQuestion;
   'assessment-instance': ResLocals &
     ResLocalsCourseInstance &
     ResLocalsAssessment &