+++ packages/formatter/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @prairielearn/formatter
 
+## 2.3.1
+
+### Patch Changes
+
+- 382dbd8: Bump dependencies
+
+## 2.3.0
+
+### Minor Changes
+
+- 07dfbca: Display "start of day"/"end of day" labels for day boundary times in `formatDateFriendly`
+- ac8e58a: Add new `formatDateISO` function
+
+### Patch Changes
+
+- b6e03e9: Upgrade dependencies
+
 ## 2.2.2
 
 ### Patch Changes