+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/logger
 
+## 3.1.3
+
+### Patch Changes
+
+- 382dbd8: Bump dependencies
+
+## 3.1.2
+
+### Patch Changes
+
+- b6e03e9: Upgrade dependencies
+
 ## 3.1.1
 
 ### Patch Changes